shill: wimax: Clear EAP passphrase after explicitly disconnecting service.
There is a workaround in WiMaxService where it requests for an EAP
passphrase on reconnection after the service is explicitly disconnected.

Due to commit "shill: make setting a property to its current value a no-op"
(bebf1b8bce52b88c2cc2d93200b9405f9c19cf21),
WiMaxService::OnEapCredentialsChanged() is no longer invoked if the
updated EAP credentials are the same as the original values, which
incorrectly keeps the service in the non-connectable.

This CL changes WiMaxService to clear the EAP passphrase when the
service is explicitly disconnected in order to ensure that
OnEapCredentialsChanged() is invoked and the service becomes connectable
again.

BUG=chromium:237945
TEST=Tested the following:
1. Build and run unit tests.
2. Manually test on a daisy with WiMAX module:
   - Connect to WiMAX.
   - Disconnect from WiMAX through UI.
   - Reconnect to WiMAX using the same credentials.

Change-Id: I0fad5c14f4d5101da417e2a0f8d5112b3f7152f4
Reviewed-on: https://gerrit.chromium.org/gerrit/50075
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>