shill_proxy: add support for matching against ServiceCompleteList
Companion to CL:50327, which adds a client-side test for the
instantiation of services from the default profile. Because a
client-side test can't assume that APs are live OTA, we need the
ability to find a service in the complete list of services, rather
than the list of visible services.

BUG=chromium:237588
TEST=network_DefaultProfileServices.py

Change-Id: I618ae3cb82595bf6ad61ed83b894398ab130b775
Reviewed-on: https://gerrit.chromium.org/gerrit/50328
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>