shill: fix crash in IPConfig::UpdateProperties
If UpdateProperties reports failure to |update_callback_|, the
device owning the IPConfig object may drop its reference to
the IPConfig. We'll then crash, when we try to EmitChanges.

Avoid the crash by having UpdateProperties hold a reference
to the IPConfig object for the duration of the method call.

BUG=chromium:237576
TEST=new unit test

Change-Id: Ia3ce1a4729429ec75d4aab6b4c84b67f573d9187
Reviewed-on: https://gerrit.chromium.org/gerrit/50509
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>