power : Add allow_docked_mode preference
Add this preference so we have a way to restore the old
behavior that suspended every time the lid is closed.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chromium:239523
TEST=Added a unit test DisallowDockedMode.
Tested this on Lumpy and Parrot. Set the allow_docked_mode
config to 0, and plug in an external display. Notice that the
system suspends instead of going to docked mode.

Change-Id: I3ab5eaf7e48b988fc9ac752d87d34743c47d6124
Reviewed-on: https://gerrit.chromium.org/gerrit/50917
Commit-Queue: Benson Leung <bleung@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>