shill: OpenVPNDriver: Work around netmask in remote
In some OpenVPN configurations, the netmask is sent in the
"ifconfig_remote" argument instead of "ifconfig_netmask" where it
should be.  Work around this by examining the content of the address.
Precedent for this processing can be found in other connection
managers, for example
https://mail.gnome.org/archives/commits-list/2009-July/msg03401.html

BUG=chromium:241264
TEST=Unit tests

Change-Id: I15c9746e8af528c357eec023327a6375e8dc2599
Reviewed-on: https://gerrit.chromium.org/gerrit/51529
Commit-Queue: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>