shill: Replaces scan_pending_ with more scan states.  Enhances output.
As part of this CL, OnAfterResume is modified to cancel any on-going scan
and restart fast scan attempts.  This change was made to in this CL to
rectify issues raised when fixing unittests as a result of this change.
This change was made at all because it was recognized that a scan
(which, with progressive scan, is now a multi-state thing) should not
take place across an arbitrarily long suspend-resume cycle.

BUG=chromium:245814
TEST=unittest and manual.  For manual, on the chromebook:

  a) shell
  b) sudo bash
  c) stop shill
  d) shill --log-level=-10 --log-scopes=wifi
  e) look in /var/log/net.log for "Scan state:"  There should be an
     entry that looks like:

     wlan0: Scan state: IDLE -> (anything) @ 0ms in to scan.

Change-Id: I1c849e42737910016bdb2c72f6f0d092150902da
Reviewed-on: https://gerrit.chromium.org/gerrit/57274
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Wade Guthrie <wdg@chromium.org>
Reviewed-by: Wade Guthrie <wdg@chromium.org>
Tested-by: Wade Guthrie <wdg@chromium.org>