power: Avoid suspending while shutting down.
This fixes an issue where closing the lid after the "clean
shutdown" interval (where other processes are given a chance
to clean up) completes but before the system powers off
results in the system suspending (only to resume and shut
down immediately when the lid is opened later).

The root cause is that the Daemon class checks its
clean_shutdown_initiated_ member to see if it's safe to
suspend, but the member was getting unset when other
processes reported readiness or the interval timed out. This
change makes Daemon's tracking of its shutdown progress
clearer and adds an additional check to the Suspender class.

BUG=chromium:245983
TEST=added unittest for Suspender change

Change-Id: I7d855c15eef04461f16cf87850199827660d4dfa
Reviewed-on: https://gerrit.chromium.org/gerrit/57302
Commit-Queue: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>