power: Turn all displays on at startup.
This fixes an issue where the displays could be stuck in the
off state if powerd was restarted after it had turned the
displays off due to inactivity.  InternalBacklightController
attempted to turn the displays on in its Init() method but
avoided sending the request due to a member variable being
previously initialized to the all-on state in the
constructor.

BUG=chromium:245083
TEST=added a unit test; also manually tested by waiting for
     displays to get turned off and then restarting powerd

Change-Id: I07fb798c7cc173b098ea2d63d1fabaf320ff7a52
Reviewed-on: https://gerrit.chromium.org/gerrit/57554
Commit-Queue: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>