Allow Chrome to set audio threads to realtime priority of 10.
Chrome's audio threads don't require much CPU, but they have to meet
short scheduling deadlines.  Audio buffers must be filled in as little
as a couple of milliseconds.  Setting these threads to real time
allows these deadlines to be met on ChromeOS.

Set the nice limit back to where it was before allowing advantageous
nice levels.

BUG=chromium:213118,chromium:221207
TEST=play audio in the background on a heavily loaded system.  Modify
Chrome to set audio threads to realtime and notice it is much better,
even with smaller buffer sizes.

Note this was already reviewed at:
https://gerrit.chromium.org/gerrit/#/c/51494/

Change-Id: Ie2dd5a15fc3e0bdf906b95fa2a050ffb06e2dfc1
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/57493
Reviewed-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>