power: Give up on broken ambient light sensors after 10 sec.
This makes the InternalBacklightController class stop
waiting for the ambient light sensor if the sensor still
hasn't supplied a reading after ten seconds. Otherwise, if
the ALS driver is broken (which has happened several times),
powerd is left in a state where it refuses to change the
backlight level in response to user inactivity.

It also adds a new Clock class that can be used to mock out
the current time. This replaces a bunch of similar code that
was spread across the power manager.

BUG=chromium:243904
TEST=added a new unittest; also manually checked that state
     changes (e.g. dimming) take effect on a device with a
     broken ambient light sensor driver

Change-Id: I818ffb74ac1839fb6689c2d943d504327f499717
Reviewed-on: https://gerrit.chromium.org/gerrit/57790
Commit-Queue: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>