Set CKA_LABEL attribute for importing X509 certificates via p11_replay.
X509 certificates will only be shown under "Your Certificates" in the
Certificate Manager if both of the following two conditions are met:
 (a) a corresponding private key is imported.
 (b) the CKA_LABEL attribute is set, and set to a non-empty string.
If either of these conditions are not met, the certificate is displayed under
"Others" in the Certificate Manager.

Note: The corresponding private key does not need to have the CKA_LABEL
attribute to be set, nor does it have to match.

This label is set in different ways based on the type of certificate
(see chromium:202503 and chromium:237870). How the label is set doesn't
actually affect where the certificate is displayed, so for testing purposes
we just set this to "testing_cert".

BUG=chromium:196315
TEST=Built image and successfully imported certificate via p11_replay. Certificate is correctly displayed under "Your Certificates" in the Certificate Manager.

Change-Id: I0aec63aa20dd8118011ce102336ebb82d82d7011
Reviewed-on: https://gerrit.chromium.org/gerrit/57681
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Tested-by: Sarah Harvey <saharvey@chromium.org>
Commit-Queue: Sarah Harvey <saharvey@chromium.org>