power: Update battery-full logic and lower threshold to 97%.
This updates powerd to report a battery as full if its
charge is greater than or equal to 97% of the
kernel-reported full charge, regardless of the current.
Previously, batteries were considered full if their charge
matched the full charge or if they were greater than or
equal to 98% with a zero current.

It also removes tricky code that pinned the battery level at
100% for a minute if the charger was removed while the
battery was reported as full.

BUG=chromium:247355
TEST=updated unit tests; also manually tested that battery
     percentages displayed in UI are continuous when
     plugging and unplugging charger while nearly full

Change-Id: Ia0d902125c9342c9069e726784ac9018f593bddc
Reviewed-on: https://gerrit.chromium.org/gerrit/57998
Commit-Queue: Daniel Erat <derat@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>