shill: build virtual_device_unittest.o for shill_unittest
Previously, it was a dependency of shill. That meant that
a) we were wasting time compiling a unit test when we just
   wanted the daemon
b) we never ran the test, because it was linked to shill_unittest
   as a library. because none of the code in shill_unittest
   referenced it, it was discarded.

Props to @davidjames for reporting a), and to @gauravsh for
helping me figure out b).

BUG=chromium:249545
TEST=manual (see below)

Manual test
-----------
- make clean
- cros_workon_make --board=link --reconf shill
- find . -name virtual_device_unittest.o
  (no match)
- cros_workon_make --board=link --test --reconf shill
- find . -name virtual_device_unittest.o
  (matches)
- GTEST_FILTER="VirtualDeviceTest.*" \
    cros_workon_make --board=link --test --reconf shill
  (runs one unit test)

Change-Id: I53ee69cf940c83ea309036d22d96ead39ecadfac
Reviewed-on: https://gerrit.chromium.org/gerrit/58612
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>