shill: cellular: Always report that SIM lock is enabled if SIM is locked.
The Cellular.SIMLockStatus/0/enabled property might be set to false, even
though the SIM is locked. This is because the corresponding ModemManager
property 'EnabledFacilityLocks' belongs to the 3GPP interface and the 3GPP
interface is not available while the Modem is in the 'LOCKED' state.

BUG=chromium:249570
TEST=(1) Build and run unit tests.
     (2) On a device running ModemManager:
          - Enable SIM lock.
          - Reset the modem.
          - Use "connectivity show devices" to verify that
            Cellular.*LockEnabled is "true".

Change-Id: Ibfdad757b5bac7da01e8387dc529559113be310c
Reviewed-on: https://gerrit.chromium.org/gerrit/58627
Reviewed-by: Arman Uguray <armansito@chromium.org>
Tested-by: Arman Uguray <armansito@chromium.org>
Commit-Queue: Arman Uguray <armansito@chromium.org>