power: Rework battery time calculations.
This updates powerd to keep a rolling average of battery
current readings rather than a rolling average of
time-to-empty and time-to-full estimates.

It also makes it report time-to-full when the battery is
charging and time-to-empty when it's discharging, rather
than time-to-full whenever a charger is connected (it's
possible for the battery to discharge when using a
low-current charger). Estimates that would be extremely
large due to a low current are reported as -1.

BUG=chromium:252948,chromium:233348,chrome-os-partner:20343
TEST=updated existing unit tests; lots of manual testing on
     different devices

Change-Id: Ic43efe9bc31832d3b6323f619a14f410d5642a0e
Reviewed-on: https://gerrit.chromium.org/gerrit/60839
Reviewed-by: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>
Commit-Queue: Daniel Erat <derat@chromium.org>