Added new pattern to post install for legacy
Because we are changing the format of UUIDs to confrom with
the linux kernel usage, PARTUUID=%U[/PARNTOFF=%d], we need
to change the post installer for legacy to fill in the uuid
correctly.

Right now, the post installer for legacy doesn't hangle the
kern_guid, so a pattern isn't needed for it.

BUG=chromium:224066
TEST=Installed and ran on Dell Latitude, ran cbuildbot

Change-Id: I44cf29a410abaf1e53cf2a349bea8a9f4255ec7a
Reviewed-on: https://gerrit.chromium.org/gerrit/59571
Tested-by: Paul Taysom <taysom@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Paul Taysom <taysom@chromium.org>