InstallAttributesIsReady: Wait for Attestation blob preparation
On some hardware, attestation blob preparation may take a noticably
longer time. In those cases, a user may have completely all the UI steps
needed to trigger an enterprise enrollment.  While enterprise enrollment
waits on TPM ownership, it does not wait for the TPM to be free and
clear.  It appears that in some instances, it is possible to create
enough TPM contention to delay a response to
InstallAttributesFinalize().

This change avoids the possibility of that contention by forcing Chrome
to serialize enrollment commitment with Attestation preparation.  Chrome
will wait for up to 10 minutes for InstallAttributesIsReady() to return
true.  This change relies on that logic to keep Chrome from making a
TPM-contention InstallAttributesFinalize call while the TPM is busy.

At the very least, this change will remove one variable from the DBus
timeout behavior we see sometimes with InstallAttributesFinalize.

TEST=Tested on ToT lumpy through dev-switch to clear then boot.
BUG=chromium:189681

Change-Id: I7adc871ac7dbb60e90f3d131318384ae42f8e074
Reviewed-on: https://gerrit.chromium.org/gerrit/59707
Reviewed-by: Will Drewry <wad@chromium.org>
Tested-by: Will Drewry <wad@chromium.org>
Commit-Queue: Will Drewry <wad@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>