Only report non-zero MiB counts
More precisely, only send an UMA event to the following UMA metrics

 Installer.{Successful,Total}MBsDownloadedFrom{Http,Https}Server

if the mechanism (the so-called DownloadSource) was actually
used. Without this patch, we would be reporting a lot of zeroes which
would dominate the histograms and thus, render them less useful than
they could be.

BUG=chromium:253612
TEST=Unittests pass

Change-Id: Ib2d797e9c65eb480e5b97e561495abffcbbfbb5d
Reviewed-on: https://gerrit.chromium.org/gerrit/59828
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: David Zeuthen <zeuthen@chromium.org>
Tested-by: David Zeuthen <zeuthen@chromium.org>