shill: Makes scan-type message a LOG(INFO)
As progressive scan gets rolled-out, it'll be really useful to know
whether a bug was seen in the presence of progressive or full scan.
This CL modifies the logging such that every shill-instigated scan is
logged with the scan type at the LOG(INFO) level.

BUG=None
TEST=unittest

Change-Id: I19763c84f5b4dda6a8a9daba056bfe1096064ccd
Reviewed-on: https://gerrit.chromium.org/gerrit/59992
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Wade Guthrie <wdg@chromium.org>
Reviewed-by: Wade Guthrie <wdg@chromium.org>
Tested-by: Wade Guthrie <wdg@chromium.org>