shill: L2TPIpsecDriver: Use PEM arrays for CACertPEM
Switch the "CACertPEM" option to use a string array of PEM.
strongSwan is equally capable of accepting PEM and DER
certificates in its CA certificate store.  Since it's not easy
to export multiple certificates in a single DER file, use
a PEM file instead of DER for the output file in the CaCertPEM
case.  Since this is the last user of CreateDERFromString(),
remove this method from CertificateFile.  Since we no longer
generate DER, there is no longer a need for CertificateFile
to hold a GLib instance.

CQ-DEPEND=CL:60542
BUG=chromium:249363
TEST=Unit tests.

Change-Id: I9acde1f165006458ea0e1cd34e5037fad4784feb
Reviewed-on: https://gerrit.chromium.org/gerrit/60557
Reviewed-by: Darin Petkov <petkov@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>