shill: WiFiService: Track successive WPA/EAP auth failures
When a WiFi association fails during 802.1x EAP authentication or
WPA 4-way handshake, track these in in the WiFi Service in a manner
that will allow us to detect multiple consecutive failures.  When
a number of such failures occur, signal the service's failure in a
similar manner to an initial authentication failure with no prior
successful connections.  This will signal the UI to notify the user
that their credentials might no longer be valid.

This counter is non-persistent (not stored to the profile) and
orthogonal to other facilities like connection throttling and
the PassphraseRequired property, since it is only meant as a
heuristic signal to Chrome.

BUG=chromium:217033
TEST=Unit tests

Change-Id: I9c2df6770e643162ae0e7b5e03b2b857f4773b37
Reviewed-on: https://gerrit.chromium.org/gerrit/61355
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>