Make setgoodkernel work with both EFI and coreboot
The chromeos-setgoodkernel script is used at the end of an
update cycle to mark the booted partition "successful" in
the GUID Partition Table (GPT).  The previous version of
this script was hard-coded to extract the booted partition
from a coreboot-style kernel invocation command-line.  For
boot from UEFI bios, the kernel command line specification
of the booted partition has a slightly different syntax.
This was causing the chromeos-setgoodkernel script to fail
and terminate without doing the "successful" marking.  Since
the autotest procedure looks for this success marker, all
autotests were failing on EFI-boot systems.

This change makes the chromeos-setgoodkernel script work for
both EFI-boot systems and coreboot systems.

BUG=none
TEST=/usr/sbin/chromeos-setgoodkernel ; cgpt show /dev/sda

Change-Id: Idba102db5c1b76235e0b183361b926190aa151c5
Reviewed-on: https://gerrit.chromium.org/gerrit/63044
Tested-by: paul drews <paul.drews@intel.com>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Josh Triplett <josh@joshtriplett.org>