shill: WiFi: Detect WEP passphrase failures
Keep track of the number of (successfully decrypted) packets we
have received on a WEP network.  If the DHCP process fails and
we have received no other packets, we should suspect that this
is due to a faulty passphrase.

BUG=chromium:216250
TEST=Unit tests + manual: Connect to a WEP network with the wrong
passphrase.  (Note that CrOS_WEP_trendnet_tew639gr is currently
refusing association for a reason other than passphrase, or is
configured for shared-key authentication which is baroque even
for WEP.)

Change-Id: I0bb3ac9539bad0cf8cfaf9115b550a631a26170b
Reviewed-on: https://gerrit.chromium.org/gerrit/61704
Commit-Queue: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>